Make some changes to improve the Windows build of DLLs for this branch
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/branches/activemq-cpp-3.4.x@1095726 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/util/concurrent/ThreadPoolExecutor.cpp
@@ -604,8 +604,8 @@ void ThreadPoolExecutor::setKeepAliveTime(long long timeout, const TimeUnit& uni
             "Keep Alive Time must be set to a non-zero value when allowCoreThreadsTimeout is enabled.");
     }
 
-    long keepAliveTime = unit.toMillis(timeout);
-    long delta = keepAliveTime - this->kernel->keepAliveTime;
+    long long keepAliveTime = unit.toMillis(timeout);
+    long long delta = keepAliveTime - this->kernel->keepAliveTime;
     this->kernel->keepAliveTime = keepAliveTime;
     if (delta < 0) {
         // TODO - When Threads are interruptible wake works so they can check timeout.
